fix(state): widen sqlc LIMIT params to int64 (CodeQL CWE-190)
The PR #2 aggregate queries used 'limit $N' with a sqlc-emitted int32
Params field. The handlers bridged from int (handler-side capped at
ObsAdmin*LimitMax = 500) via int32(...) which CodeQL flagged as
possible value truncation (int -> int32 on a value that has no bounds
check at the cast site).

sqlc's postgres engine doesn't honor @arg macros for LIMIT/OFFSET
(GitHub sqlc/sqlc#3840), so the canonical fix is to cast the SQL
parameter to int8 in queries.sql. sqlc then emits Column2/Column3
int64 Params fields. The handlers widen int -> int64 (safe; no loss)
and the pgstore/memstore pass the int64 straight through to lib/pq,
which transmits it as a Postgres bigint (the protocol-side cap is
2^63-1, well above the SQL-level 2^31-1 limit ceiling).

Pin rationale inlined next to each LIMIT clause so future readers see
why the cast is there rather than reaching for @arg.

+++ pkg/state/queries.sql
@@ -610,7 +610,9 @@ where consumed_at is null
 --                      (handler default: now() - 24h, hard cap 168h)
 --   * $2 baseline    — RFC 3339 lower bound for the baseline pool
 --                      (handler default: now() - 7d, fixed by ADR)
---   * $3 limit       — top-N by deviation (handler default 50, cap 200)
+--   * $3 limit       — top-N by deviation (handler default 50, cap 200;
+--                      cast to int8 so sqlc emits int64 Params and the
+--                      handler's int→int64 widening is safe)
 -- Result columns:
 --   * account_id, app_id, minute, current_mb_seconds
 --   * baseline_mean, baseline_stddev, baseline_samples
@@ -683,7 +685,7 @@ select account_id,
 from scored
 where z_score is not null
 order by z_score desc
-limit $3;
+limit $3::int8;
 
 -- name: PerAccountRateLimitAggregate :many
 -- ADR-091 §3.5 — operator observability backend (PR #2) durable view.
@@ -691,7 +693,9 @@ limit $3;
 -- window, grouped by subject (account_id, NULL for anonymous actors).
 --   * $1 since  — RFC 3339 lower bound (handler default: now() - 24h,
 --                 hard cap 168h per pkg/api/limits.go::ObsAdminWindowMaxHours)
---   * $2 limit  — top-N by hits (handler default 100, cap 500)
+--   * $2 limit  — top-N by hits (handler default 100, cap 500;
+--                 cast to int8 so sqlc emits int64 Params and the
+--                 handler's int→int64 widening is safe)
 -- Anonymous (subject IS NULL) rows are bucketed under a single
 -- account_id = NULL row so the operator UI can render the "anon
 -- credential stuffing" signal distinctly from named-account bursts.
@@ -706,5 +710,5 @@ where kind = 'auth.rate_limited'
   and at >= $1
 group by coalesce(subject, '00000000-0000-0000-0000-000000000000'::uuid)
 order by hits desc, last_event_at desc
-limit $2;
+limit $2::int8;
